Separate analytics function into separate file
